refactor: split export.rs into focused modules (types, stats, snapshot, timestamp) (#980)
Split the monolithic src/export.rs (778 lines) into a src/export/ module
directory with four focused sub-modules. All public API re-exported from
mod.rs for backward compatibility. No functional changes.
